requirements.txt
boto3==1.34.11
PyYAML==5.3.1
ansible
